fix(error): don't silent git error in get_diff
the git error from call to `checkout_last_commit_at_paths` shouldn't be silenced.
Fix : #1960
If the git repository files are corrupted, we need to have an error.

I have tried to reproduce a case where we can have an error from
`checkout_last_commit_at_paths` because of no commit to include in diff,
but without success, to get the `there are no commits` as valid message.