fix: More generous retry for failed HTTP/2 registry request
With some private registries that do not support HTTP/2, or perhaps falsely advertise as supporting it, reqwest does not always identify the cause of failure as a connection error. To work around this, make our retry with HTTP/1.1 more generous by retrying any error that is not clearly caused by a status code returned by the server.