fix: Retry HTTP/2 GOAWAY from registry with HTTP/1.1
With some private registries that do not support HTTP/2, or perhaps falsely advertise as supporting it, reqwest does not always identify the cause of failure as a connection error. In this case, the underlying `h2` library may return a premature `GOAWAY` error from either the client or server. If this happens, we should use HTTP/1.1 instead.

This PR implements this by walking up the chain of errors to see if a "go away" error is a cause. This is a fairly explicit check, but the downside is that we have to add `h2` as a direct dependency in order to pull in the `h2::Error` type. (It is already a transitive dependency through `reqwest`.)

Related to #1668 and #2261.